+++ lib/cyclium/migrations.ex
@@ -40,7 +40,8 @@ defmodule Cyclium.Migrations do
     21 => Cyclium.Migrations.V21,
     22 => Cyclium.Migrations.V22,
     23 => Cyclium.Migrations.V23,
-    24 => Cyclium.Migrations.V24
+    24 => Cyclium.Migrations.V24,
+    25 => Cyclium.Migrations.V25
   }
 
   @doc "All known migration versions, ascending. Use to run every version in order."